wrap ApiClient with TileDBClient
