+++ .openapi-generator-ignore
@@ -29,5 +29,5 @@ README.md
 settings.gradle
 
 #the api client contains the custom code for static initialization from the json file
-src/main/java/io/tiledb/cloud/rest_api/ApiClient.java
+src/main/java/io/tiledb/cloud/TileDBClient.java
 src/main/java/io/tiledb/cloud/rest_api/Login.java
